Backend testing okay
Have a more or less stable backend for the parser,
there is a set of functions to create a set of voiced pitch segments
given a pitch selector. At this point a pitch selector is a somewhat
amorphous concept that is guiding the development process. It's
currently being represented as some form of query that is used to
select pitches from an array of pitch segments.

The way the module is currently working, the program could basically
form a pool of voiced pitches given the data segment of pitch classes
and an array of octave and voicing directions to use to build the
initial query.

Some further testing of the function parse_pitch_attrs should reveal
more about how the frontend might work.