Slight cleanup
As I started to mention with the last commit, at this point it is
possible to use the marana module in a way that produces output for use
in lilypond scores.

The type of workflow meant here is demonstrated in the examples
directory. Each example is essentially a single script that is focused
on creating one voice of the score.

The actual arrangement of the voices within the lilypond score is
something that I'm quite happy to handle in lilypond format, as this
tends to give me a little more intuitive control of what is happening
musically speaking (or muscially musicing).