Added a makefile to build the examples
Renamed some files and created a simple makefile with a single score.pdf
target. This commit provides a very simple example of what the proposed
end-to-end workflow might look like.

Basically, some of the lessons that I learned from the last time I was
building a score like this were to:
1. include a makefile
2. try to keep the lilypond files legible as music (rather than very
   abstract collections of variables and references)