Removing source from morch branch
Discovered git-worktree which makes it possible to more clearly separate
data and logic. can now include the morch branch in a subfolder which
makes our workflow a bit easier. It's possible add new features at the
level of the main branch and try them out on a score within the morch
subfolder.