Restructured build process
Consolodate functions used repeatedly to generate score
segments by moving the functions up to ../../src/marana/tools.py.

The score segments in this context refer to .ly files. In this current
commit, they are being compiled and collected in the sketches directory.
In a later step, they will be linked into the scores.