Rudimentary working build
Makefile updated to generate score and parts. The segments will need to
be updated manually in the score and parts each time that we generate
a new section for the piece. This is probably a sub-optimal solution,
but it was just proving too tricky to figure out a workflow that
utilizes lilypond tags.

Anyway, the most important part is that this commit represents a full
end-to-end version of the piece that is basically ready for road. A few
things still need to be added, in the way of rehearsal marks etc. But
for now this should be enought to be getting on with.