Add branch handler to printer for generic lilypond
This makes it possible to use some of the same utility functions for
printing music expressions to stdout. Previously, we had been relying on
wrapping all objects to be printed in some sort of abjad based classes.
This is not necessary (i.e. pretty dumb and overkill) for most use cases
that involve a score that is being designed from the bottom up.