Commit first "complete" score build
This commit unifies the previously seperate sections, laying

  stringcanon
  wwindChorale
  iyghBrass

end-to-end. Also adding a todo list for the changes that need to take
place over the next couple of days. In terms of additions from the
workshop, the next batch of things to be done are the changes to the
brass segment.