Add near completions to first segment
This articulates a musical evolution through the whole first segment,
bringing in all instruments that we are interested in hearing. There may
be some slight edits left to do with the horns -> they appear to have a
range that is slightly stretched.

Otherwise, it's looking pretty good for the first segment.