Final updates before v1 release
Cleaned up build script to fully automate the distribution (includes
now a small wrapper for the build directory that can be easily
archived. Final edits in all the four sections as well as the
score.ly. Inclusion of a new file dist.org that basically generates
the cover for the score in latex.